requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
discord.py==1.6.0
flake8==3.8.4
idna==2.10
mccabe==0.6.1
multidict==5.1.0
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.15.0
requests==2.25.1
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.6.3
